+++ Runtime/Core/TweenManager.cs
@@ -8,7 +8,7 @@ using Object = UnityEngine.Object;
 namespace FlowTween {
     
 /// <summary>
-/// Singleton for running and pooling <see cref="TweenBase"/> objects at runtime.
+/// Singleton for running and pooling <see cref="Runnable"/> objects at runtime.
 /// </summary>
 public class TweenManager : MonoBehaviour {
     [SerializeField] bool _enabled = true;
@@ -40,8 +40,8 @@ public class TweenManager : MonoBehaviour {
         }
     }
 
-    readonly Dictionary<Type, Queue<TweenBase>> _inactive = new();
-    readonly List<TweenInstance> _active = new();
+    readonly Dictionary<Type, Queue<Runnable>> _inactive = new();
+    readonly List<RunnableInstance> _active = new();
 
     /// <summary>
     /// Whether or not to do null checks on
@@ -88,7 +88,7 @@ public class TweenManager : MonoBehaviour {
     void Update() {
         for (var i = 0; i < _active.Count; i++) {
             var instance = _active[i];
-            var tween = instance.Tween;
+            var tween = instance.Runnable;
 
             var isNull = DoNullChecks && instance.HasOwner && instance.Owner == null;
             if (!isNull) {
@@ -103,10 +103,10 @@ public class TweenManager : MonoBehaviour {
         }
     }
 
-    void Return(TweenBase tween) {
+    void Return(Runnable tween) {
         var type = tween.GetType();
         if (!_inactive.TryGetValue(type, out var queue)) {
-            _inactive.Add(type, queue = new Queue<TweenBase>());
+            _inactive.Add(type, queue = new Queue<Runnable>());
         }
 
         queue.Enqueue(tween);
@@ -116,10 +116,10 @@ public class TweenManager : MonoBehaviour {
     /// Cancels a tween.
     /// </summary>
     /// <param name="tween">The tween to cancel.</param>
-    /// <param name="callOnComplete">Whether or not to call the tween's <see cref="TweenBase.CompleteAction"/></param>
-    public void Cancel(TweenBase tween, bool callOnComplete = false) {
+    /// <param name="callOnComplete">Whether or not to call the tween's <see cref="Runnable.CompleteAction"/></param>
+    public void Cancel(Runnable tween, bool callOnComplete = false) {
         if (callOnComplete) tween.Cancel(false);
-        _active.RemoveAll(t => t.Tween == tween);
+        _active.RemoveAll(t => t.Runnable == tween);
         Return(tween);
     }
 
@@ -130,13 +130,13 @@ public class TweenManager : MonoBehaviour {
     /// The owner of the tweens to cancel.
     /// Any tweens owned by components on this are also cancelled.
     /// </param>
-    /// <param name="callOnComplete">Whether or not to call the tweens' <see cref="TweenBase.CompleteAction"/></param>
+    /// <param name="callOnComplete">Whether or not to call the tweens' <see cref="Runnable.CompleteAction"/></param>
     public void CancelObject(GameObject obj, bool callOnComplete = false) {
         for (var i = 0; i < _active.Count; i++) {
             var instance = _active[i];
             if (instance.Owner != obj) continue;
 
-            Cancel(instance.Tween, callOnComplete);
+            Cancel(instance.Runnable, callOnComplete);
             i--;
         }
     }
@@ -156,11 +156,11 @@ public class TweenManager : MonoBehaviour {
     /// will be used as the owner.
     /// Otherwise, the tween will have no owner.
     /// </param>
-    /// <typeparam name="T">A subclass of <see cref="TweenBase"/> to get.</typeparam>
+    /// <typeparam name="T">A subclass of <see cref="Runnable"/> to get.</typeparam>
     /// <exception cref="InvalidOperationException">
     /// The maximum number of tweens has been reached (see <see cref="MaxTweens"/>).
     /// </exception>
-    public T Get<T>(Object owner = null) where T : TweenBase, new() {
+    public T Get<T>(Object owner = null) where T : Runnable, new() {
         var gameObjectOwner = owner switch {
             GameObject o => o,
             Component c => c.gameObject,
@@ -181,7 +181,7 @@ public class TweenManager : MonoBehaviour {
             tween = new T();
         }
 
-        _active.Add(TweenInstance.From(tween, gameObjectOwner));
+        _active.Add(RunnableInstance.From(tween, gameObjectOwner));
         return tween;
     }
     
@@ -200,6 +200,8 @@ public class TweenManager : MonoBehaviour {
     /// </param>
     /// <typeparam name="T">The type of the tween to get.</typeparam>
     public Tween<T> NewTween<T>(Object owner = null) => Get<Tween<T>>(owner);
+    
+    public Sequence NewSequence(Object owner = null) => Get<Sequence>(owner);
     
     /// <summary>
     /// Tries to get the singleton instance and run the given action on it.
@@ -227,19 +229,19 @@ public class TweenManager : MonoBehaviour {
         return true;
     }
 
-    readonly struct TweenInstance {
-        public readonly TweenBase Tween;
+    readonly struct RunnableInstance {
+        public readonly Runnable Runnable;
         public readonly GameObject Owner;
         public readonly bool HasOwner;
 
-        public TweenInstance(TweenBase tween, GameObject owner, bool hasOwner) {
-            Tween = tween;
+        public RunnableInstance(Runnable runnable, GameObject owner, bool hasOwner) {
+            Runnable = runnable;
             Owner = owner;
             HasOwner = hasOwner;
         }
         
-        public static TweenInstance From(TweenBase tween, GameObject owner) {
-            return new TweenInstance(tween, owner, owner != null);
+        public static RunnableInstance From(Runnable tween, GameObject owner) {
+            return new RunnableInstance(tween, owner, owner != null);
         }
     }
 }