+++ Runtime/Templates/TweenSettingsProperty.cs
@@ -8,6 +8,10 @@ namespace FlowTween.Templates {
 /// shared scriptable object <see cref="TweenSettingsTemplateAsset}"/>.
 /// </summary>
 [Serializable]
-public class TweenSettingsProperty : TemplateProperty<TweenSettings> { }
+public class TweenSettingsProperty : TemplateProperty<TweenSettings> {
+    public TweenSettingsProperty() { }
+    public TweenSettingsProperty(TweenSettings inlineValue) : base(inlineValue) { }
+    public TweenSettingsProperty(TemplateAsset<TweenSettings> templateValue) : base(templateValue) { }
+}
 
 }